Separate iteration from functionality
