+++ src/hello_lib.h
@@ -1,0 +1,7 @@
+#ifndef HELLO_LIB_H
+#define HELLO_LIB_H
+
+void blink (int n);
+char invert_capitalization (char c);
+
+#endif
